+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 
+## [0.3.2] - 2018-08-24
+### Added
+- DOM.nodeName() & .selectors()
+
+### Fixed
+- Namespacing `<svg>` elements in DOM.createElement()
+
+
 ## [0.3.1] - 2018-08-20
 ### Added
 - DOM.dropFiles()
@@ -33,7 +41,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial boilerplate code.
 
 
-[Unreleased]: https://github.com/Sphinxxxx/abo-utils/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/Sphinxxxx/abo-utils/compare/v0.3.2...HEAD
+[0.3.2]:      https://github.com/Sphinxxxx/abo-utils/compare/v0.3.1...v0.3.2
 [0.3.1]:      https://github.com/Sphinxxxx/abo-utils/compare/v0.3.0...v0.3.1
 [0.3.0]:      https://github.com/Sphinxxxx/abo-utils/compare/v0.2.0...v0.3.0
 [0.2.0]:      https://github.com/Sphinxxxx/abo-utils/compare/v0.1.0...v0.2.0